Move to a services project so it is more testable
