+++ src/STO.Models/Interfaces/IPlayerService.cs
@@ -1,4 +1,4 @@
-namespace STO.Wasm.Interfaces;
+namespace STO.Models.Interfaces;
 
 /// <summary>
 /// Service for working with Players. Virtual objects made from PlayerEntity with calculated properties from other entities
@@ -6,9 +6,9 @@ namespace STO.Wasm.Interfaces;
 public interface IPlayerService
 {
     /// <summary>
-    /// Gets a Player from RowKey
+    /// Gets a Player based on Id
     /// </summary>
-    /// <param name="rowKey">The RowKey for the PlayerEntity to match on</param>
+    /// <param name="id">The Id to match on</param>
     /// <returns>A Player.</returns>
-    public Player GetPlayer(string rowKey);
+    public Player GetPlayer(string id);
 }
