+++ src/STO.Wasm/Pages/Admin.razor
@@ -3,6 +3,7 @@
 @inject IConfiguration Configuration
 @inject IWebAssemblyHostEnvironment Env
 @using System.Security.Claims
+@using STO.Models.Interfaces
 @inject AuthenticationStateProvider AuthenticationStateProvider
 @inject IDataService DataService
 @inject IGameEntityService GameEntityService
