+++ src/STO.Wasm/Pages/PlayerDetailsTransactions.razor
@@ -1,4 +1,5 @@
 @page "/players/{PlayerUrlSegment}/transactions"
+@using STO.Models.Interfaces
 @inject IDataService DataService
 @inject IPlayerEntityService PlayerEntityService
 @inject ITransactionEntityService TransactionEntityService
