+++ src/STO.Wasm/Services/PlayerEntityService.cs
@@ -1,4 +1,5 @@
 using System.Globalization;
+using STO.Models.Interfaces;
 
 namespace STO.Wasm.Services
 {
