+++ src/STO.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "STO.Wasm", "STO.Wasm\STO.Wa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "STO.DataManager", "STO.DataManager\STO.DataManager.csproj", "{EC2C85C3-A54D-453B-9AE4-C9803C84AD0B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "STO.Services", "STO.Services\STO.Services.csproj", "{3ECCC708-0629-499B-9B64-2BFB0F2A101B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -69,6 +71,18 @@ Global
 		{EC2C85C3-A54D-453B-9AE4-C9803C84AD0B}.Release|x64.Build.0 = Release|Any CPU
 		{EC2C85C3-A54D-453B-9AE4-C9803C84AD0B}.Release|x86.ActiveCfg = Release|Any CPU
 		{EC2C85C3-A54D-453B-9AE4-C9803C84AD0B}.Release|x86.Build.0 = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|x64.Build.0 = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Debug|x86.Build.0 = Debug|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|x64.ActiveCfg = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|x64.Build.0 = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|x86.ActiveCfg = Release|Any CPU
+		{3ECCC708-0629-499B-9B64-2BFB0F2A101B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
